AOC2023: Day 5

+++ 2023/aoc2023/src/main/java/org/michaelrol/Day.java
@@ -7,8 +7,8 @@ package org.michaelrol;
 
 public interface Day {
 
-  public abstract int Part1();
+  public abstract long Part1();
 
-  public abstract int Part2();
+  public abstract long Part2();
 
 }
